Added an IE specific rule in layout CSS to prevent command icons hidden behind HTMLArea generated tables
When inserting table in HTMLArea, if the table is set to align to left or right, IE will make the table float so the command button will be hidden behind it.

So a IE specific rule is added in layout CSS to set command buttons clear on both side. This will apply to all themes based on the standard layout. And other browsers are not affected/needed.

Merged from MOODLE_15_STABLE