Added some new settings to theme config.php so that a file meta.php can be automatically included into the HEAD section of the current theme.
Supports parent and standard files too.

I added this to support the inclusion of some IE specific CSS  :-( which needs
to be done from HTML and not CSS, (MDL-7782) but it could be used to include other things
as well, or even to run some PHP code, set various vriables etc without messing
up the theme header.html (and regardless of the theme being used).